requirements.txt
fastapi
strawberry-graphql
uvicorn
sqlalchemy
asyncpg
python-dotenv
